backend/pyproject.toml
[project]
name = "be1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "matplotlib>=3.10.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.2",
]
